// src/generated/api.ts
export interface operations {
  bookkeepingSystemVersion: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** Successful Request */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          /**
           * @example
           *   {
           *   "objects": {
           *   "version": "2.0"
           *   }
           *   }
           */
          "application/json": {
            objects?: {
              /** @enum {string} */
              version?: "1.0" | "2.0";
            };
          };
        };
      };
      /** Authentication required */
      401: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
// ...
}